+++ client/src/utils/api/mailingList.ts
@@ -1,8 +1,8 @@
 import { MailingListRequest } from 'types/mailingList';
 import { AuthResponse, Method } from 'types/network';
-import { makeRequest } from 'utils/network/request';
-import { APIRoutes } from 'utils/network/endpoints';
-import * as M from 'utils/network/errorMessages';
+import { makeRequest } from 'utils/api/request';
+import { APIRoutes } from 'utils/api/endpoints';
+import * as M from 'utils/api/errorMessages';
 
 export const mailingListSubmit = async (data: MailingListRequest) => {
   return await makeRequest<AuthResponse, MailingListRequest>(
