+++ client/src/utils/api/request.ts
@@ -1,23 +1,13 @@
 import axios from 'axios';
 import { Method } from 'types/network';
 
-const DEV_URL = 'http://localhost:5000';
-
-export const getUrl = (urlExt: string) => {
-  if (process.env.NODE_ENV === 'production') {
-    return urlExt;
-  }
-  return DEV_URL + urlExt;
-};
-
 export const makeRequest = async <R, D = Record<string, unknown>>(
   method: Method,
-  urlExt: string,
+  route: string,
   errorMessage: string,
   data?: D,
   token?: string,
 ): Promise<R> => {
-  const url = getUrl(urlExt);
   const options = {
     headers: {
       Authorization: 'Bearer ' + token,
@@ -27,13 +17,13 @@ export const makeRequest = async <R, D = Record<string, unknown>>(
   try {
     switch (method) {
       case Method.GET:
-        return (await (await axios.get(url, options)).data) as R;
+        return (await (await axios.get(route, options)).data) as R;
       case Method.POST:
-        return (await (await axios.post(url, data, options)).data) as R;
+        return (await (await axios.post(route, data, options)).data) as R;
       case Method.PATCH:
-        return (await (await axios.patch(url, data, options)).data) as R;
+        return (await (await axios.patch(route, data, options)).data) as R;
       case Method.DELETE:
-        return (await (await axios.delete(url, options)).data) as R;
+        return (await (await axios.delete(route, options)).data) as R;
       default:
         return {} as R;
     }
